+++ boltprojection/adaptor.go
@@ -2,20 +2,21 @@ package boltprojection
 
 import (
 	"context"
+	"time"
 
 	"github.com/dogmatiq/dogma"
 	"github.com/dogmatiq/projectionkit/internal/identity"
 	"github.com/dogmatiq/projectionkit/internal/unboundhandler"
+	"github.com/dogmatiq/projectionkit/resource"
 	"go.etcd.io/bbolt"
 )
 
 // adaptor adapts a boltprojection.ProjectionMessageHandler to the
 // dogma.ProjectionMessageHandler interface.
 type adaptor struct {
-	MessageHandler
-
-	db  *bbolt.DB
-	key string
+	db      *bbolt.DB
+	handler MessageHandler
+	repo    *ResourceRepository
 }
 
 // New returns a new Dogma projection message handler by binding a
@@ -32,10 +33,19 @@ func New(
 	}
 
 	return &adaptor{
-		MessageHandler: h,
-		db:             db,
-		key:            identity.Key(h),
+		db:      db,
+		handler: h,
+		repo: NewResourceRepository(
+			db,
+			identity.Key(h),
+		),
 	}
+}
+
+// Configure produces a configuration for this handler by calling methods on
+// the configurer c.
+func (a *adaptor) Configure(c dogma.ProjectionConfigurer) {
+	a.handler.Configure(c)
 }
 
 // HandleEvent updates the projection to reflect the occurrence of an event.
@@ -45,81 +55,39 @@ func (a *adaptor) HandleEvent(
 	s dogma.ProjectionEventScope,
 	m dogma.Message,
 ) (bool, error) {
-	tx, err := a.db.Begin(true)
-	if err != nil {
-		// CODE COVERAGE: This branch can not be easily covered without somehow
-		// breaking the BoltDB connection or the database file in some way.
-		return false, err
-	}
-	defer tx.Rollback() // nolint:errcheck
-
-	ok, err := updateVersion(ctx, tx, a.key, r, c, n)
-	if !ok || err != nil {
-		return ok, err
-	}
-
-	if err := a.MessageHandler.HandleEvent(ctx, tx, s, m); err != nil {
-		return false, err
-	}
-
-	return true, tx.Commit()
+	return a.repo.UpdateResourceVersionFn(
+		ctx,
+		r, c, n,
+		func(ctx context.Context, tx *bbolt.Tx) error {
+			return a.handler.HandleEvent(ctx, tx, s, m)
+		},
+	)
 }
 
 // ResourceVersion returns the version of the resource r.
 func (a *adaptor) ResourceVersion(ctx context.Context, r []byte) ([]byte, error) {
-	return queryVersion(ctx, a.db, a.key, r)
+	return a.repo.ResourceVersion(ctx, r)
 }
 
 // CloseResource informs the projection that the resource r will not be
 // used in any future calls to HandleEvent().
 func (a *adaptor) CloseResource(ctx context.Context, r []byte) error {
-	return deleteResource(ctx, a.db, a.key, r)
+	return a.repo.DeleteResource(ctx, r)
 }
 
-// StoreResourceVersion sets the version of the resource r to v
-func (a *adaptor) StoreResourceVersion(ctx context.Context, r, v []byte) error {
-	tx, err := a.db.Begin(true)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback() // nolint:errcheck
-
-	if err := storeVersion(ctx, tx, a.key, r, v); err != nil {
-		return err
-	}
-
-	return tx.Commit()
-}
-
-// UpdateResourceVersion updates the version of the resource r to n without
-// handling any event.
-//
-// If c is not the current version of r, it returns false and no update occurs.
-func (a *adaptor) UpdateResourceVersion(
-	ctx context.Context,
-	r, c, n []byte,
-) (ok bool, err error) {
-	tx, err := a.db.Begin(true)
-	if err != nil {
-		return false, err
-	}
-	defer tx.Rollback() // nolint:errcheck
-
-	ok, err = updateVersion(ctx, tx, a.key, r, c, n)
-	if !ok || err != nil {
-		return ok, err
-	}
-
-	return true, tx.Commit()
-}
-
-// DeleteResource removes all information about the resource r from the
-// handler's data store.
-func (a *adaptor) DeleteResource(ctx context.Context, r []byte) error {
-	return deleteResource(ctx, a.db, a.key, r)
+// TimeoutHint returns a duration that is suitable for computing a deadline
+// for the handling of the given message by this handler.
+func (a *adaptor) TimeoutHint(m dogma.Message) time.Duration {
+	return a.handler.TimeoutHint(m)
 }
 
 // Compact reduces the size of the projection's data.
 func (a *adaptor) Compact(ctx context.Context, s dogma.ProjectionCompactScope) error {
-	return a.MessageHandler.Compact(ctx, a.db, s)
+	return a.handler.Compact(ctx, a.db, s)
+}
+
+// ResourceRepository returns a repository that can be used to manipulate the
+// handler's resource versions.
+func (a *adaptor) ResourceRepository() resource.Repository {
+	return a.repo
 }