+++ CHANGELOG.md
@@ -9,6 +9,14 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Add `resource.Repository` interface
+- Add `boltprojection.ResourceRepository`
+- Add `sqlprojection.ResourceRepository`
+
 ## [0.6.3] - 2021-01-29
 
 ### Fixed